Move to more performant kafka python client
confluent-kafka (maintained by Confluent, the maintainers of kafka)
manages to vastly outperform our previously library through the use of
the librdkafka C library. We see a 5x increase in connector performance
compared the previous library (with a slight additional optimization in
the form of the linger.ms variable). This should give plenty of headroom
for receiving position messages.

I also added the flight's id as the key for each kafka message, making
it easy to add more partitions down the line if we for some reason
desire additional kafka performance. We're currently CPU bound in both
the connector and db-updater so I don't see that being necessary yet.