Add postgres client for convenient inspection of db
This would allow someone to inspect the contents of the positions table
(or the flights table if they're using Postgres for that) from inside
the db-updater container. An alternative option would be to externally
expose Postgres' port on the TimescaleDB container so that it can be
accessed from outside of the internal Compose network by a tool of the
user's choice.