Specify optional docker-compose variables, ensure kafka is ready (#40)
Need to explicitly tell docker-compose that we want these variables
provided to the container. If they exist in the host environment,
they'll be passed to the container. If not, they won't be.
This ensures that the environment variables are correctly picked
up by the python app if set.

Set dependencies on docker-compose services to start up in the right order

Ensure that kafka is ready to receive events from producers before
kicking off the connector. Otherwise kafka may be in a transient state
that errors out.

Run CI on all commits and allow manual runs